+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 8.2.0-dev
 - added option to disable .phpt tests in mytester.php with --noPhpt
+- deprecated TestCase::METHOD_PATTERN
 
 Version 8.1.1
 - fixed error in mytester.php when package konecnyjakub/phpt-runner is not installed